Count of Smaller Numbers After Self
You are given an integer array nums and you have to return a new counts array. The counts array has the property where counts[i] is the number of smaller elements to the right of nums[i].

 

Example 1:

Input: nums = [5,2,6,1]
Output: [2,1,1,0]
Explanation:
To the right of 5 there are 2 smaller elements (2 and 1).
To the right of 2 there is only 1 smaller element (1).
To the right of 6 there is 1 smaller element (1).
To the right of 1 there is 0 smaller element.

Example 2:

Input: nums = [-1]
Output: [0]

Example 3:

Input: nums = [-1,-1]
Output: [0,0]

 

Constraints:

    1 <= nums.length <= 105
    -104 <= nums[i] <= 104

Start from the last number and keep numbers in BST (in python it is SortedList). Then when we see new number we need to find how many elements in this list smaller than this number, and then put this number into our sorted list.

Complexity

Complexity of both operations : add to list and do binary search with bisect_left is O(log n), so overall complexity is O(n log n).

Remark

In this type of problems there is often solutions using segment tree or BIT: the idea is to for each number get its order statistics, that is what will be the place if we sort. For example, if we have 3,1,4,5,2, then we create [0,0,0,0,0] array, then we meet number 3 first, so we update it like [0,0,1,0,0], then we have [1,0,1,0,0], then we have [1,0,1,1,0], [1,0,1,1,1], [1,1,1,1,1]. Instead of keeping list, what we keep is Segment Tree or BIT for quick answers for ranges: actually what we want to evaluate on each step is sum of elements in some range [0, ind]. Complexity of these approaches will be O(n log n) as well.


In Short, Binary indexed Tree is useful for Updating an Array Element in O(logn) Time and Calculating the prefix sum of an Array in O(logn) Time.

Question : Given an array arr, For every Array Index i, count the No. of numbers less than arr[i] to the right side of index i.

Approach:
Since we require the elements count to the right side of index i, let's traverse the array in the reverse order and store the count of elements as we traverse through them.

Let's consider the array , nums = [5,2,6,1].

We need a counter array to keep track of the count of elements we encountered till the ith index. For this example, the maximum element is 6. So we need a counter array of size 6 to track the elements count.

For index i=3

indices				0 1 2 3 4 5 6
counter array (cnt) 0 0 0 0 0 0 0
i=3 => arr[3]=1 => we need No. of numbers less than 1. we can Get this count from the counter array i.e, cnt[0]=0
				   Now update the counter array by increasing the count of arr[3]

indices				0 1 2 3 4 5 6
counter array (cnt) 0 1 0 0 0 0 0

For index i=2:

indices				0 1 2 3 4 5 6
counter array (cnt) 0 1 0 0 0 0 0
i=2 => arr[2]=6 => we need No. of numbers less than 6. we can Get this count from the counter array i.e, cnt[0]+cnt[1]+cnt[2]+cnt[3]+cnt[4]+cnt[5]=1
				   Now update the counter array by increasing the count of 6.

indices				0 1 2 3 4 5 6
counter array (cnt) 0 1 0 0 0 0 1

For index i=1:

indices				0 1 2 3 4 5 6
counter array (cnt) 0 1 0 0 0 0 1
i=1 => arr[1]=2 => we need No. of numbers less than 2. we can Get this count from the counter array i.e, cnt[0]+cnt[1]=1
				   Now update the counter array by increasing the count of 2.

indices				0 1 2 3 4 5 6
counter array (cnt) 0 1 1 0 0 0 1

For index i=0:

indices				0 1 2 3 4 5 6
counter array (cnt) 0 1 1 0 0 0 1
i=0 => arr[0]=5 => we need No. of numbers less than 5. we can Get this count from the counter array i.e, cnt[0]+cnt[1]+cnt[2]+cnt[3]+cnt[4]=2
				   Now update the counter array by increasing the count of 5.

indices				0 1 2 3 4 5 6
counter array (cnt) 0 1 1 0 0 1 1

Note: The numbers can be negative. So, add a offset to convert negative numbers to positive numbers.
-10^4 <= nums[i] <= 10^4
Let's consider a offset of 10^4+1. updated range of elements will be 1<=nums[i]+offset <= 2*10^4+1
Now we can store the negative elements count as well.

Optimization:
=> Traversing the counter array for every index to fetch the count of elements smaller than arr[i] is time consuming - O(n)
=> Use the Binary Indexed Tree for the counter array, it will query the counter array prefix sums in O(logn) and it will update the counter array (increase the count of elements) in O(logn).

Time Complexity:

=> MAX -> maximum element in the array i.e 2*10^4+1 after adding the offset.
=> O(n * logMAX * logMAX) => for every index i (0<=i<n) , we have to query for count(logMAX) and update the element count(logMAX)