Lowest Common Ancestor of a Binary Tree
Given a binary tree, find the lowest common ancestor (LCA) of two given nodes in the tree.

According to the definition of LCA on Wikipedia: “The lowest common ancestor is defined between two nodes p and q as the lowest node in T that has both p and q as descendants (where we allow a node to be a descendant of itself).”

 

Example 1:

Input: root = [3,5,1,6,2,0,8,null,null,7,4], p = 5, q = 1
Output: 3
Explanation: The LCA of nodes 5 and 1 is 3.

Example 2:

Input: root = [3,5,1,6,2,0,8,null,null,7,4], p = 5, q = 4
Output: 5
Explanation: The LCA of nodes 5 and 4 is 5, since a node can be a descendant of itself according to the LCA definition.

Example 3:

Input: root = [1,2], p = 1, q = 2
Output: 1

 

Constraints:

    The number of nodes in the tree is in the range [2, 105].
    -109 <= Node.val <= 109
    All Node.val are unique.
    p != q
    p and q will exist in the tree.

The idea is to traverse our tree and keep auxilary function dfs(node), which returns 0 if subtree do not have desired nodes and 1 if it has (one or two of them). For given node, we calculate information about:

    lft is dfs(node.left), that is answer for left subtree.
    rgh is dfs(node.right), that is answer for right subtree.
    mid is true if current node is either p or q.

Then what we need to find is such node, for which at least 2 out of these 3 values are true. Why? If p is parent of q, then the answer is p and we have mid = 1 and one of the lft and rgh equal to 1. If q is parent of p, it is similar. If we have h which is LCA of p and q, then for h we have lft = rgh = 1 and mid = 0.

Complexity

Complexity is O(n), both time and space.