Decode Ways II
A message containing letters from A-Z can be encoded into numbers using the following mapping:

'A' -> "1"
'B' -> "2"
...
'Z' -> "26"

To decode an encoded message, all the digits must be grouped then mapped back into letters using the reverse of the mapping above (there may be multiple ways). For example, "11106" can be mapped into:

    "AAJF" with the grouping (1 1 10 6)
    "KJF" with the grouping (11 10 6)

Note that the grouping (1 11 06) is invalid because "06" cannot be mapped into 'F' since "6" is different from "06".

In addition to the mapping above, an encoded message may contain the '*' character, which can represent any digit from '1' to '9' ('0' is excluded). For example, the encoded message "1*" may represent any of the encoded messages "11", "12", "13", "14", "15", "16", "17", "18", or "19". Decoding "1*" is equivalent to decoding any of the encoded messages it can represent.

Given a string s containing digits and the '*' character, return the number of ways to decode it.

Since the answer may be very large, return it modulo 109 + 7.

 

Example 1:

Input: s = "*"
Output: 9
Explanation: The encoded message can represent any of the encoded messages "1", "2", "3", "4", "5", "6", "7", "8", or "9".
Each of these can be decoded to the strings "A", "B", "C", "D", "E", "F", "G", "H", and "I" respectively.
Hence, there are a total of 9 ways to decode "*".

Example 2:

Input: s = "1*"
Output: 18
Explanation: The encoded message can represent any of the encoded messages "11", "12", "13", "14", "15", "16", "17", "18", or "19".
Each of these encoded messages have 2 ways to be decoded (e.g. "11" can be decoded to "AA" or "K").
Hence, there are a total of 9 * 2 = 18 ways to decode "1*".

Example 3:

Input: s = "2*"
Output: 15
Explanation: The encoded message can represent any of the encoded messages "21", "22", "23", "24", "25", "26", "27", "28", or "29".
"21", "22", "23", "24", "25", and "26" have 2 ways of being decoded, but "27", "28", and "29" only have 1 way.
Hence, there are a total of (6 * 2) + (3 * 1) = 12 + 3 = 15 ways to decode "2*".

 

Constraints:

    1 <= s.length <= 105
    s[i] is a digit or '*'.

Classical dynamic programming problem, where we need to carefully connect dp_new with dp. There is more clean solution, if we keep 3 states: dp[0] is number of total ways,dp[1] is number of decodes, ending with 1, which is not used yet and dp[2] is number of decodes, ending with 2, which is not used yet. Then we can do the following recursion:

    If c == "*" then there can be two options: if last number has only one digit, then we have 9*dp[0] options. Also, if we try go group last two digits, we have 9*dp[1] options for ...1* case and 6*dp[2] options for ...2* case. To evaluate new dp[1] and dp[2], we just return dp[0].
    If c != "*" then total number of options is (c > '0') * dp[0] for single digit last number or we could pair it with 1 or with 2 if it is not more than 6. To evaluate new dp[1], we check if last digit is 1 and for dp[2], we check if last digit is 2.

Complexity

Time complexity is O(n), space complexity is O(1).