Push Dominoes
There are n dominoes in a line, and we place each domino vertically upright. In the beginning, we simultaneously push some of the dominoes either to the left or to the right.

After each second, each domino that is falling to the left pushes the adjacent domino on the left. Similarly, the dominoes falling to the right push their adjacent dominoes standing on the right.

When a vertical domino has dominoes falling on it from both sides, it stays still due to the balance of the forces.

For the purposes of this question, we will consider that a falling domino expends no additional force to a falling or already fallen domino.

You are given a string dominoes representing the initial state where:

    dominoes[i] = 'L', if the ith domino has been pushed to the left,
    dominoes[i] = 'R', if the ith domino has been pushed to the right, and
    dominoes[i] = '.', if the ith domino has not been pushed.

Return a string representing the final state.

 

Example 1:

Input: dominoes = "RR.L"
Output: "RR.L"
Explanation: The first domino expends no additional force on the second domino.

Example 2:
  https://s3-lc-upload.s3.amazonaws.com/uploads/2018/05/18/domino.png
Input: dominoes = ".L.R...LR..L.."
Output: "LL.RR.LLRRLL.."


Explanation:
First, we iterate the dominoes twice.

    From left to right, we want to check which dominoes might be falling to the right side.
    So every time we see a 'R' we know that the next dominoes will maybe fall right, so we save in the 'right' vector the distance from the previous 'R'. (Until we reach a 'L').
    From right to left, we do the same for the left side.
    Example:
    Let's use the second example from the question - ".L.R...LR..L..".
    The 'right' array will look like this: [0,0,0,0,1,2,3,0,0,1,2,0,0,0]
    The 'left' array will look like this: [1,0,0,0,3,2,1,0,0,2,1,0,0,0]

Next step is to compare 'right and 'left'.
For each i:

    If both are 0 - we just copy what we had in dominoes.
    If right[i] is 0 but left[i] is not, it means this domino falls to the left, and vice versa.
    if they are both equal, which means that we have balance of the forces and we stay vertical.
    If we have to different values in right[i] and left[i], we find the minimum which means that it's closer.

 

Constraints:

    n == dominoes.length
    1 <= n <= 105
    dominoes[i] is either 'L', 'R', or '.'.