Word Ladder II
A transformation sequence from word beginWord to word endWord using a dictionary wordList is a sequence of words beginWord -> s1 -> s2 -> ... -> sk such that:

    Every adjacent pair of words differs by a single letter.
    Every si for 1 <= i <= k is in wordList. Note that beginWord does not need to be in wordList.
    sk == endWord

Given two words, beginWord and endWord, and a dictionary wordList, return all the shortest transformation sequences from beginWord to endWord, or an empty list if no such sequence exists. Each sequence should be returned as a list of the words [beginWord, s1, s2, ..., sk].

 

Example 1:

Input: beginWord = "hit", endWord = "cog", wordList = ["hot","dot","dog","lot","log","cog"]
Output: [["hit","hot","dot","dog","cog"],["hit","hot","lot","log","cog"]]
Explanation: There are 2 shortest transformation sequences:
"hit" -> "hot" -> "dot" -> "dog" -> "cog"
"hit" -> "hot" -> "lot" -> "log" -> "cog"

Example 2:

Input: beginWord = "hit", endWord = "cog", wordList = ["hot","dot","dog","lot","log"]
Output: []
Explanation: The endWord "cog" is not in wordList, therefore there is no valid transformation sequence.

 

Constraints:

    1 <= beginWord.length <= 5
    endWord.length == beginWord.length
    1 <= wordList.length <= 1000
    wordList[i].length == beginWord.length
    beginWord, endWord, and wordList[i] consist of lowercase English letters.
    beginWord != endWord
    All the words in wordList are unique.

Approach: we are trying to build out a tree in a breadth-first fashion. We start the root of the tree with beginWord and then use a queue to populate it's children. The filter used for finding the children words of that node is word belongs to wordList and word is 1 letter away from the parent. At the end of every level, we make sure that we mark all the nodes as visited so we don't form a cycle in the tree generation.

If we build this tree out level by level in a breadth first fashion, we might find the end word on a particular level. Since we are finding the minimum depth branch of the tree which led to the endWord, all levels below this in the tree can be simply discarded, however, we should still continue to finish checking the whole level to see if there is any other branch of the tree on the same level which has endWord. Once we finish that, we can stop the breadth-first traversal and return the branches.

To implement this approach, I use a queue to do a BFS of the tree, where as the tree nodes themselves have a pointer to their parent. Once we have located the nodes with endWord, we can traverse the branch upto the root using the parent pointers to generate the output.

Note: This is probably not the most optimal or concise solution, but it was easy to explain and visualize. If you can suggest any improvements to this, please leave it below in the comments.