Non-negative Integers without Consecutive Ones
Given a positive integer n, return the number of the integers in the range [0, n] whose binary representations do not contain consecutive ones.

 

Example 1:

Input: n = 5
Output: 5
Explanation:
Here are the non-negative integers <= 5 with their corresponding binary representations:
0 : 0
1 : 1
2 : 10
3 : 11
4 : 100
5 : 101
Among them, only integer 3 disobeys the rule (two consecutive ones) and the other 5 satisfy the rule. 

Example 2:

Input: n = 1
Output: 2

Example 3:

Input: n = 2
Output: 3

 

Constraints:

    1 <= n <= 109

Let us consider n = 10100101101, then what numbers we can choose:

0XXXXXXXXXX,
100XXXXXXXX,
101000XXXXX,
10100100XXX,
101001010XX,

where X can be any digit: 0 or 1. So, we need to found number of sequences with 0 and 1 without consecutive ones, and it is Fibonacci numbers. So, we use binary representation of number and traverse it from the start and check if we have two consecutive ones. If we have, we break, if not, we add corresponding Fibonacci number.

Complexity

Time complexity is O(32), because we use binary representation of number. If we have unbounded n, complexity is O(m) = O(log n), where m is number of bits in this number (and if we need to give answer modulo some N to find Fibonacci numbers fast).