3Sum Closest
Given an array nums of n integers and an integer target, find three integers in nums such that the sum is closest to target. Return the sum of the three integers. You may assume that each input would have exactly one solution.

 

Example 1:

Input: nums = [-1,2,1,-4], target = 1
Output: 2
Explanation: The sum that is closest to the target is 2. (-1 + 2 + 1 = 2).

 

Constraints:

    3 <= nums.length <= 10^3
    -10^3 <= nums[i] <= 10^3
    -10^4 <= target <= 10^4

Idea

    Sort nums in increasing order.
    Let's fix nums[i] by iterating i in range [0..n-2], we using 2 pointers to find 2 elements in range [i+1..n-1], so that nums[i] + nums[l] + nums[r] will have minimum difference with our target.