Beautiful Array
For some fixed n, an array nums is beautiful if it is a permutation of the integers 1, 2, ..., n, such that:

For every i < j, there is no k with i < k < j such that nums[k] * 2 = nums[i] + nums[j].

Given n, return any beautiful array nums.  (It is guaranteed that one exists.)

 

Example 1:

Input: n = 4
Output: [2,1,4,3]

Example 2:

Input: n = 5
Output: [3,1,2,5,4]

 

Note:

    1 <= n <= 1000

 
Seriously, this question should be categorized as hard to those who have no experience of ACM, such like me.
however, there still is a way to come across the solution without mathematical intuition and reasoning skills, which needs a pen and a paper, of coz, and a little bit observation.

First, it is not hard to figure out that,
odd + even = odd != 2 * x
so devide the array from 1 to N into two parts (take N=10 as example),
1 3 5 7 9 | 2 4 6 8 10
then the question becomes how to make the left or the right be "beautiful array".
look at the right part that includes even numbers, here the pen and paper are in need.
say we have 2 ,4, where should we put 6? apprently not the right most position, while the left most slot is ok, so we get 6, 2, 4.
likewisely, we get 6, 2, 4, 8, and it is beautiful too. Given this, it seems that we can rebuild the left part spirally, but can it work always?
Unfortunately, the answer is no. 10, 6, 2, 4, 8 is not beautiful.
here comes the KEY -- if we splitted the spiral array into two parts and rebuild each part spirally again, we got two beautiful arrays,
10 2 6 | 4 8, AND, any number in the sub left plus any number in the sub right is a double of odd, while all numbers in this array are even.
but are we to the end? 18 10 2 6 14 is not beautiful once more. so split and rebuild, 18 2 10 | 6 14, (left x + right y) / 2 is not in the array itself.
it can be easily verified that all the above steps are also feasible on original odd subarrays.
having gone so far, it is not difficult to think of devide and conquer and recursion, and TRY.

Theory

Take extra note of : A[k] * 2 = A[i] + A[j] Notice how A[k] * 2 is always even!
So now we just have to make sure that A[i] + A[j] is always odd, and we can do that by making left odd and right even.
Effectivley its a divide and conquer algorithm where we split the array into two halves, a left half which is odd, and a right half which is even.
Eventually we have to put back both the halves together and we multiply the odd portion by 2 * i - 1 and the even portion with 2 * i. We have to do this step to return the larger array from our "current" beautiful array - why? - TBH I don't even know :/ But do watch this youtube video I have linked.

Credit

https://www.youtube.com/watch?v=O-1ucu8ErEo