Trapping Rain Water
Given n non-negative integers representing an elevation map where the width of each bar is 1, compute how much water it can trap after raining.

 

Example 1:
https://assets.leetcode.com/uploads/2018/10/22/rainwatertrap.png
Input: height = [0,1,0,2,1,0,1,3,2,1,2,1]
Output: 6
Explanation: The above elevation map (black section) is represented by array [0,1,0,2,1,0,1,3,2,1,2,1]. In this case, 6 units of rain water (blue section) are being trapped.

Example 2:

Input: height = [4,2,0,3,2,5]
Output: 9

 

Constraints:

    n == height.length
    0 <= n <= 3 * 104
    0 <= height[i] <= 105

To solve this problem, we need to understand what does trapping means. Imagine that we have some index i, how to evaluate amount of water we have in this index? There are two constraints: water can go to the left and to the right. Imagine for the moment, that we have infinite wall in the left, than what is the level of water it can rise? It is the maximum value from our map to the right of our index! The same logic works if we add infinite wall in the left. What to do now if we do not have infninite walls? We need to find minimum between the highest point to the left and the highest point to the right and this will solve the problem!

Complexity

We have O(n) time and memory complexity.