Stone Game
Alex and Lee play a game with piles of stones.  There are an even number of piles arranged in a row, and each pile has a positive integer number of stones piles[i].

The objective of the game is to end with the most stones.  The total number of stones is odd, so there are no ties.

Alex and Lee take turns, with Alex starting first.  Each turn, a player takes the entire pile of stones from either the beginning or the end of the row.  This continues until there are no more piles left, at which point the person with the most stones wins.

Assuming Alex and Lee play optimally, return True if and only if Alex wins the game.

 

Example 1:

Input: piles = [5,3,4,5]
Output: true
Explanation: 
Alex starts first, and can only take the first 5 or the last 5.
Say he takes the first 5, so that the row becomes [3, 4, 5].
If Lee takes 3, then the board is [4, 5], and Alex takes 5 to win with 10 points.
If Lee takes the last 5, then the board is [3, 4], and Alex takes 4 to win with 9 points.
This demonstrated that taking the first 5 was a winning move for Alex, so we return true.

 

Constraints:

    2 <= piles.length <= 500
    piles.length is even.
    1 <= piles[i] <= 500
    sum(piles) is odd.

Solution 1: Top down DP

    Please note that both players are playing optimally, so we just play optimally regardless of who is Alex, who is Lee.
    Let dp(left, right) return [firstScore, secondScore] where firstScore is the maximum score when the player1 play first, secondScore is the maximum score when the player2 play second, they pick stones in piles[left]...piles[right].
    For stones in piles[left]...piles[right], there are 2 choices for the player1 to pick:
        Pick left: pickLeft = dp(left + 1, right).
            The score of the player1 = piles[left] + the second pick score from pickLeft, so firstScore = piles[left] + pickLeft[1]
            The score of the player2 = first pick score from pickLeft, so secondScore = pickLeft[0]
        Pick right: pickRight = dp(left, right - 1).
            The score of the player1 = piles[right] + the second pick score from pickRight, so firstScore = piles[right] + pickRight[1]
            The score of the player2 = first pick score from pickRight, so secondScore = pickRight[0].
    We need to get the maximum score when the player1 plays first from above 2 choices.
    Finally, dp(0, len(piles) - 1) return [firstScore, secondScore], where alexScore = firstScore since Alex plays first, leeScore = secondScore since Lee plays second.
Complexity

    Time: O(N^2), where N <= 500 is the size of piles array.
    Space: O(N^2)

 Solution 2: Bottom up DP

    Just convert from Top down DP to Bottom up DP to eliminate recursion memory stack.
Complexity

    Time: O(N^2), where N <= 500 is the size of piles array.
    Space: O(N^2)