Minimum Window Substring
Given two strings s and t of lengths m and n respectively, return the minimum window substring of s such that every character in t (including duplicates) is included in the window. If there is no such substring, return the empty string "".

The testcases will be generated such that the answer is unique.

A substring is a contiguous sequence of characters within the string.

 

Example 1:

Input: s = "ADOBECODEBANC", t = "ABC"
Output: "BANC"
Explanation: The minimum window substring "BANC" includes 'A', 'B', and 'C' from string t.

Example 2:

Input: s = "a", t = "a"
Output: "a"
Explanation: The entire string s is the minimum window.

Example 3:

Input: s = "a", t = "aa"
Output: ""
Explanation: Both 'a's from t must be included in the window.
Since the largest window of s only has one 'a', return empty string.

 

Constraints:

    m == s.length
    n == t.length
    1 <= m, n <= 105
    s and t consist of uppercase and lowercase English letters.

 
Follow up: Could you find an algorithm that runs in O(m + n) time?
   Hide Hint #1  
Use two pointers to create a window of letters in S, which would have all the characters from T.
   Hide Hint #2  
Since you have to find the minimum window in S which has all the characters from T, you need to expand and contract the window using the two pointers and keep checking the window for all the characters. This approach is also called Sliding Window Approach.

L ------------------------ R , Suppose this is the window that contains all characters of T 
                          
        L----------------- R , this is the contracted window. We found a smaller window that still contains all the characters in T

When the window is no longer valid, start expanding again using the right pointer. 

The idea of sliding window with 2 pointers: we create counter cnt_t is frequencies how many time s we need to take each symbol, for example for abca we have a:2, b:1, c:1. We create also cnt_s as empty counter, which will keep information about frequencies is current window [beg, end) - not that we do not include end. To quickly understand how many symbols have good frequency, we have found variable. Now, we can have two options to change our window:

    If found == len(cnt_t), it means that we have in our window all symbols we want: no need to extend it to the right: we will shrink it to the left. We need to update our ans, where we keep pair (end - beg, beg) , then we look at the first element old = s[beg] and if we have cnt_s[old] == cnt_t[old] it means that we have exactly the right frequency for this symbol. So, when we move beg one position to the right, we need to decrease found by one. Finally, we decrease frequency by one and morebeg` to the right.
    In the opposite case, first we check if end == len(s) and if it is true, we break. Then we look at the new symbol new = s[end] (because it was not included, windows are [beg, end). We check if cnt_s[new] == cnt_t[new] - 1 and if it is the case, it means, that when we add new symbol we need to increase found by one. Then we increase frequency and move end one position to the right.

Complexity

Time complexity is O(m + n), becuse on each step we move beg or end one point to the right. Space complexity is O(A), where A is the size of alphabet.