Count Good Nodes in Binary Tree
Given a binary tree root, a node X in the tree is named good if in the path from root to X there are no nodes with a value greater than X.

Return the number of good nodes in the binary tree.

 

Example 1:

Input: root = [3,1,4,3,null,1,5]
Output: 4
Explanation: Nodes in blue are good.
Root Node (3) is always a good node.
Node 4 -> (3,4) is the maximum value in the path starting from the root.
Node 5 -> (3,4,5) is the maximum value in the path
Node 3 -> (3,1,3) is the maximum value in the path.

Example 2:

Input: root = [3,3,null,4,2]
Output: 3
Explanation: Node 2 -> (3, 3, 2) is not good, because "3" is higher than it.

Example 3:

Input: root = [1]
Output: 1
Explanation: Root is considered as good.

 

Constraints:

    The number of nodes in the binary tree is in the range [1, 10^5].
    Each node's value is between [-10^4, 10^4].

   Hide Hint #1  
Use DFS (Depth First Search) to traverse the tree, and constantly keep track of the current path maximum.

In this problem we need to traverse tree and find all nodes which have all ancestors less or equal than given node. Let us traverse our tree and change its nodes, such that in each node we will keep the biggest value on path node -> .. -> root.

    We update self.count if child.val >= root.val, that is node child is good.
    We update value for child, in this case when we run recursion for this node, we know the biggest value on the path between this node and root.
    Finally, we run recursion for every children.

Complexity

It is just O(n) for time and O(h) for additional space, because we use recursion.