Maximum Product of Splitted Binary Tree
Given the root of a binary tree, split the binary tree into two subtrees by removing one edge such that the product of the sums of the subtrees is maximized.

Return the maximum product of the sums of the two subtrees. Since the answer may be too large, return it modulo 109 + 7.

Note that you need to maximize the answer before taking the mod and not after taking it.

 

Example 1:
https://assets.leetcode.com/uploads/2020/01/21/sample_1_1699.png
Input: root = [1,2,3,4,5,6]
Output: 110
Explanation: Remove the red edge and get 2 binary trees with sum 11 and 10. Their product is 110 (11*10)

Example 2:
https://assets.leetcode.com/uploads/2020/01/21/sample_2_1699.png
Input: root = [1,null,2,3,4,null,null,5,6]
Output: 90
Explanation: Remove the red edge and get 2 binary trees with sum 15 and 6.Their product is 90 (15*6)

Example 3:

Input: root = [2,3,9,10,7,8,6,5,4,11,1]
Output: 1025

Example 4:

Input: root = [1,1]
Output: 1

 

Constraints:

    The number of nodes in the tree is in the range [2, 5 * 104].
    1 <= Node.val <= 104

   Hide Hint #1  
If we know the sum of a subtree, the answer is max( (total_sum - subtree_sum) * subtree_sum) in each node.

Idea

    If we remove any edge of the Binary Tree, we can forms: a sub-tree and its complement.
    Firstly, we dfs to calculate the total sum of all nodes in the Binary Tree.
    Then, we dfs in post order to calculate sum of each sub-tree in the Binary Tree.
        If the sum of current sub-tree is currSum, then its complement is totalSum - currSum.
        We update the answer if currSum * (totalSum - currSum) is greater than current answer.

https://assets.leetcode.com/users/images/e06c1df2-dc3b-4ab6-82be-85222efeb9a6_1629358867.9065757.png

Complexity

    Time: O(N), where N <= 5*10^4 is the number of nodes in the Binary Tree.
    Space: O(H), where H is the height of the Binary Tree, it's the stack depth of the recursion.
        In the best case, the binary tree is balanced, the height is O(logN)
        In the worst case, the binary tree is is a skewed binary tree, the height is O(N)