Rectangle Area II
We are given a list of (axis-aligned) rectangles. Each rectangle[i] = [xi1, yi1, xi2, yi2] , where (xi1, yi1) are the coordinates of the bottom-left corner, and (xi2, yi2) are the coordinates of the top-right corner of the ith rectangle.

Find the total area covered by all rectangles in the plane. Since the answer may be too large, return it modulo 109 + 7.

 

Example 1:
https://s3-lc-upload.s3.amazonaws.com/uploads/2018/06/06/rectangle_area_ii_pic.png
Input: rectangles = [[0,0,2,2],[1,0,2,3],[1,0,3,1]]
Output: 6
Explanation: As illustrated in the picture.

Example 2:

Input: rectangles = [[0,0,1000000000,1000000000]]
Output: 49
Explanation: The answer is 1018 modulo (109 + 7), which is (109)2 = (-7)2 = 49.

 

Constraints:

    1 <= rectangles.length <= 200
    rectanges[i].length = 4
    0 <= rectangles[i][j] <= 109
    The total area covered by all rectangles will never exceed 263 - 1 and thus will fit in a 64-bit signed integer.

In the first place, the problem is a member of a large problem genre (so-called computational geometry problems). 218. Skyline Problem and 391. Perfect Rectangle are similar instances.

The key idea of sweep line is to consider the bounds of geometric shapes in a specific direction without need to take care of the values that lie between those bounds, since their properties wouldn't change until meeting the next bound. Exploiting this observation, we usually scan the figures along x-axis and maintain certain states expressed in y-coordinates. This is generally called a sweep line. See more details in CLRS 33.2.

With this approach, we should be able to quickly construct a workable O(n^2) running time solution with the help of balanced BST. However, the fundamental difficulty to improve this upper bound is to maintain a dynamic interval structure that allows us to update and query the total length of all intervals in the current interval collection in o(n^2) time. (Little-oh notation means a strictly less upper bound.)

This can be perfectly achieved through segment tree, a data structure exactly designed for dynamic interval update and query. The basic idea in segment tree is divide-and-conquer, dividing the interval into two equal-length sub-intervals and perform update and query operations on them recursively. For range updates, we shall use a lazy / pending variable to store the pending value to apply to sub-intervals when updating or querying them.

Generally segment tree takes O(log n) per operation, where n is the size of element array length. Here we cannot construct a full segment tree for [0, 10^9], so we deploy a virtual segment tree that only creates node when necessary. The implicit segment tree still keeps its running time, whereas lowers its space complexity to log R (R is the range of possible values).

In conclusion, this problem can be solved in O(n^2) if the input size is not large. For larger inputs, O(n log R) solution would be a good choice, though I really don't expect that anyone work it out in 45 mins.